HOTFIX: Move Swipe to the components section
